% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr.R
\name{phyndr_combn}
\alias{phyndr_combn}
\title{Generate all distinct sets of tip labels from a phyndr tree}
\usage{
phyndr_combn(phy, max_n = 1000)
}
\arguments{
\item{phy}{A phyndr tree}

\item{max_n}{Maximum number of trees to generate; if the number of
distinct trees is larger than this, we'll throw an error instead.}
}
\description{
Generate all distinct sets of tip labels from a phyndr tree
}

