% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr.R
\name{phyndr_genus}
\alias{phyndr_genus}
\title{Phyndr genus}
\usage{
phyndr_genus(phy, data_species)
}
\arguments{
\item{phy}{An ape phylogeny}

\item{data_species}{A vector of species names for which we have
trait data.  Species names in both the tree and in this vector
must be separated with underscores, not with spaces.}
}
\description{
Most simple minded version of phyndr that works on genera only
}
\details{
The algorithm:

1: drop genera if there is a species match for that genus, but
don't drop the actual matches.

2: work out which genera can be collapsed to a single tip due to
monophyly.

2a. For each genus, determine if they are monophyletic

2b. Then, for genera that are not monophyletic determine which can
be made monophyletic by dropping groups that are not represented
in the data set.

2c. Collapse each genus down, dropping only genera that are
required to achieve monophyly.

3. For each collapsed node we'll mangle the name to "genus::name"
and then make a list of suitable species in the tree and data set;
for now stored as \code{clades}.
}

