% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr.R
\name{phyndr_n_distinct}
\alias{phyndr_n_distinct}
\title{Number of distinct trees}
\usage{
phyndr_n_distinct(phy)
}
\arguments{
\item{phy}{A phyndr phylogeny}
}
\description{
Number of possible distinct trees that can be generated from a
phyndr set.
}

