% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr.R
\name{phyndr_sample}
\alias{phyndr_sample}
\title{Sample one distinct set of tip labels from a phyndr tree}
\usage{
phyndr_sample(phy)
}
\arguments{
\item{phy}{A phyndr tree}
}
\description{
Sample one distinct set of tip labels from a phyndr tree
}

