% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr.R
\name{phyndr_taxonomy}
\alias{phyndr_taxonomy}
\title{Phyndr taxonomic}
\usage{
phyndr_taxonomy(phy, data_species, taxonomy)
}
\arguments{
\item{phy}{An ape phylogeny}

\item{data_species}{A vector of species names for which we have
trait data.  Species names in both the tree and in this vector
must be separated with underscores, not with spaces.}

\item{taxonomy}{A data.frame with taxonomic information.  Row
names must be present and must list every species in \code{phy}
and every species in \code{data_species}.  One or more columns
must be present; the first column is the lowest (finest) taxonomic
grouping and the last column is the highest (coarsest) taxonomic
grouping.  The names are arbitrary but will be used in creating
mangled names in the resulting phylogeny.}
}
\description{
Taxonomic method of phyndr that works with generalised sets of
taxonomic information.  Requires a nested set of taxonomic classes
(e.g, genus, family, order, etc) but does not assume that these
classes are necessarily monophyletic.
}

