% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr_topo.R
\name{phyndr_topo}
\alias{phyndr_topo}
\title{phyndr for topo trees}
\usage{
phyndr_topo(phy, data_species, topo)
}
\arguments{
\item{phy}{An ape phylogeny}

\item{data_species}{A vector of species names for which we have
trait data.  Species names in both the tree and in this vector
must be separated with underscores, not with spaces.}

\item{topo}{A topological tree used to determine patterns of
relatedness among species in \code{phy}.  The better resolved this
tree is and the better the overlap it has with \code{phy} and
\code{data_species} the more useful this will be.  Branch length
information is not used.}
}
\description{
phyndr for topo trees
}

