% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/phyndr_topology.R
\name{phyndr_topology}
\alias{phyndr_topology}
\title{phyndr for topo trees}
\usage{
phyndr_topology(phy, data_species, topology)
}
\arguments{
\item{phy}{An ape phylogeny.  The tree must be ultrametric and
this is enforced on entry.}

\item{data_species}{A vector of species names for which we have
trait data.  Species names in both the tree and in this vector
must be separated with underscores, not with spaces.}

\item{topology}{A topological tree used to determine patterns of
relatedness among species in \code{phy}.  The better resolved this
tree is and the better the overlap it has with \code{phy} and
\code{data_species} the more useful this will be.  Branch length
information is not used.}
}
\description{
phyndr for topo trees
}

